Move runtime libraries from iree/tools/ to runtime/src/iree/utils/.
