+++ CMakeLists.txt
@@ -634,6 +634,9 @@ if(IREE_BUILD_COMPILER)
   add_subdirectory(compiler)
 endif()
 
+add_subdirectory(runtime)
+
+# tools/ can depend on compiler/ and runtime/
 add_subdirectory(iree/tools)
 
 if(IREE_BUILD_TRACY)
@@ -690,14 +693,10 @@ set(IREE_PUBLIC_INCLUDE_DIRS "${IREE_COMMON_INCLUDE_DIRS}"
 add_subdirectory(build_tools/benchmarks)
 
 #-------------------------------------------------------------------------------
-# Transitional directories.
+# Samples
 #-------------------------------------------------------------------------------
 
-# The runtime uses tools and targets from elsewhere in the tree in a way
-# that is not supported by CMake, so we include it last (it aggregates
-# many things into top-level APIs). This should resolve once we finish
-# relayering everything.
-add_subdirectory(runtime)
+# samples/ can depend on anything, so we include it last
 if(IREE_BUILD_SAMPLES)
   add_subdirectory(samples)
 endif()
