+++ compiler/src/iree/compiler/ConstEval/BUILD
@@ -74,10 +74,10 @@ iree_compiler_cc_library(
     ],
     deps = [
         "//compiler/src/iree/compiler/Dialect/VM/Target/Bytecode",
-        "//iree/tools/utils:vm_util",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/vmvx/registration",
         "//runtime/src/iree/modules/hal",
+        "//runtime/src/iree/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
