+++ compiler/src/iree/compiler/ConstEval/CMakeLists.txt
@@ -67,7 +67,7 @@ iree_cc_library(
     iree::hal
     iree::hal::vmvx::registration
     iree::modules::hal
-    iree::tools::utils::vm_util
+    iree::utils::vm_util
     iree::vm
     iree::vm::bytecode_module
     iree::vm::cc
