+++ iree/tools/BUILD
@@ -17,7 +17,6 @@ cc_binary(
     name = "iree-benchmark-module",
     srcs = ["iree-benchmark-module-main.cc"],
     deps = [
-        "//iree/tools/utils:vm_util",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:cc",
         "//runtime/src/iree/base:tracing",
@@ -26,6 +25,7 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
+        "//runtime/src/iree/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -37,8 +37,6 @@ cc_binary(
     name = "iree-benchmark-trace",
     srcs = ["iree-benchmark-trace-main.c"],
     deps = [
-        "//iree/tools/utils:trace_replay",
-        "//iree/tools/utils:yaml_util",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:tracing",
         "//runtime/src/iree/base/internal:atomic_slist",
@@ -47,6 +45,8 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/testing:benchmark",
+        "//runtime/src/iree/utils:trace_replay",
+        "//runtime/src/iree/utils:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
@@ -57,7 +57,6 @@ cc_binary(
     testonly = True,
     srcs = ["iree-check-module-main.cc"],
     deps = [
-        "//iree/tools/utils:vm_util",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:cc",
         "//runtime/src/iree/base:core_headers",
@@ -70,6 +69,7 @@ cc_binary(
         "//runtime/src/iree/modules/check",
         "//runtime/src/iree/modules/hal",
         "//runtime/src/iree/testing:gtest",
+        "//runtime/src/iree/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
     ],
@@ -129,7 +129,6 @@ cc_binary(
         "//compiler/src/iree/compiler/Tools:init_passes_and_dialects",
         "//compiler/src/iree/compiler/Tools:init_targets",
         "//compiler/src/iree/compiler/Translation:IREEVM",
-        "//iree/tools/utils:vm_util",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:cc",
         "//runtime/src/iree/base:logging",
@@ -138,6 +137,7 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
+        "//runtime/src/iree/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -155,7 +155,6 @@ cc_binary(
     name = "iree-run-module",
     srcs = ["iree-run-module-main.cc"],
     deps = [
-        "//iree/tools/utils:vm_util",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:cc",
         "//runtime/src/iree/base:tracing",
@@ -164,6 +163,7 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
+        "//runtime/src/iree/utils:vm_util",
         "//runtime/src/iree/vm",
         "//runtime/src/iree/vm:bytecode_module",
         "//runtime/src/iree/vm:cc",
@@ -174,14 +174,14 @@ cc_binary(
     name = "iree-run-trace",
     srcs = ["iree-run-trace-main.c"],
     deps = [
-        "//iree/tools/utils:trace_replay",
-        "//iree/tools/utils:yaml_util",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:tracing",
         "//runtime/src/iree/base/internal:file_path",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
+        "//runtime/src/iree/utils:trace_replay",
+        "//runtime/src/iree/utils:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
@@ -191,9 +191,6 @@ cc_binary(
     name = "iree-e2e-matmul-test",
     srcs = ["iree-e2e-matmul-test.c"],
     deps = [
-        "//iree/tools/utils:cpu_features",
-        "//iree/tools/utils:trace_replay",
-        "//iree/tools/utils:yaml_util",
         "//runtime/src/iree/base",
         "//runtime/src/iree/base:core_headers",
         "//runtime/src/iree/base:tracing",
@@ -202,6 +199,9 @@ cc_binary(
         "//runtime/src/iree/hal",
         "//runtime/src/iree/hal/drivers",
         "//runtime/src/iree/modules/hal",
+        "//runtime/src/iree/utils:cpu_features",
+        "//runtime/src/iree/utils:trace_replay",
+        "//runtime/src/iree/utils:yaml_util",
         "//runtime/src/iree/vm",
         "@com_github_yaml_libyaml//:yaml",
     ],
