+++ iree/tools/CMakeLists.txt
@@ -28,7 +28,6 @@ endif()
 
 add_subdirectory(android)
 add_subdirectory(test)
-add_subdirectory(utils)
 
 iree_cc_binary(
   NAME
@@ -45,7 +44,7 @@ iree_cc_binary(
     iree::hal
     iree::hal::drivers
     iree::modules::hal
-    iree::tools::utils::vm_util
+    iree::utils::vm_util
     iree::vm
     iree::vm::bytecode_module
     iree::vm::cc
@@ -66,8 +65,8 @@ iree_cc_binary(
     iree::hal::drivers
     iree::modules::hal
     iree::testing::benchmark
-    iree::tools::utils::trace_replay
-    iree::tools::utils::yaml_util
+    iree::utils::trace_replay
+    iree::utils::yaml_util
     iree::vm
     yaml
 )
@@ -90,7 +89,7 @@ iree_cc_binary(
     iree::modules::check
     iree::modules::hal
     iree::testing::gtest
-    iree::tools::utils::vm_util
+    iree::utils::vm_util
     iree::vm
     iree::vm::bytecode_module
   TESTONLY
@@ -122,7 +121,7 @@ iree_cc_binary(
     iree::base::tracing
     iree::hal::drivers
     iree::modules::hal
-    iree::tools::utils::vm_util
+    iree::utils::vm_util
     iree::vm
     iree::vm::bytecode_module
     iree::vm::cc
@@ -141,8 +140,8 @@ iree_cc_binary(
     iree::hal
     iree::hal::drivers
     iree::modules::hal
-    iree::tools::utils::trace_replay
-    iree::tools::utils::yaml_util
+    iree::utils::trace_replay
+    iree::utils::yaml_util
     iree::vm
     yaml
 )
@@ -160,9 +159,9 @@ iree_cc_binary(
     iree::hal
     iree::hal::drivers
     iree::modules::hal
-    iree::tools::utils::cpu_features
-    iree::tools::utils::trace_replay
-    iree::tools::utils::yaml_util
+    iree::utils::cpu_features
+    iree::utils::trace_replay
+    iree::utils::yaml_util
     iree::vm
     yaml
 )
@@ -274,7 +273,7 @@ if(IREE_BUILD_COMPILER)
       iree::hal
       iree::hal::drivers
       iree::modules::hal
-      iree::tools::utils::vm_util
+      iree::utils::vm_util
       iree::vm
       iree::vm::bytecode_module
       iree::vm::cc
