+++ iree/tools/android/run_module_app/CMakeLists.txt
@@ -34,7 +34,7 @@ iree_cc_library(
     iree::base::cc
     iree::hal::drivers
     iree::modules::hal
-    iree::tools::utils::vm_util
+    iree::utils::vm_util
     iree::vm
   LINKOPTS
     "-landroid"
