+++ iree/tools/android/run_module_app/src/main.cc
@@ -15,7 +15,7 @@
 #include "iree/base/status_cc.h"
 #include "iree/hal/drivers/init.h"
 #include "iree/modules/hal/module.h"
-#include "iree/tools/utils/vm_util.h"
+#include "iree/utils/vm_util.h"
 #include "iree/vm/api.h"
 #include "iree/vm/bytecode_module.h"
