+++ iree/tools/iree-benchmark-trace-main.c
@@ -17,8 +17,8 @@
 #include "iree/hal/api.h"
 #include "iree/hal/drivers/init.h"
 #include "iree/testing/benchmark.h"
-#include "iree/tools/utils/trace_replay.h"
-#include "iree/tools/utils/yaml_util.h"
+#include "iree/utils/trace_replay.h"
+#include "iree/utils/yaml_util.h"
 #include "iree/vm/api.h"
 
 IREE_FLAG(string, driver, "vmvx", "Backend driver to use.");
