+++ iree/tools/iree-check-module-main.cc
@@ -25,7 +25,7 @@
 #include "iree/modules/hal/module.h"
 #include "iree/testing/gtest.h"
 #include "iree/testing/status_matchers.h"
-#include "iree/tools/utils/vm_util.h"
+#include "iree/utils/vm_util.h"
 #include "iree/vm/api.h"
 #include "iree/vm/bytecode_module.h"
