+++ runtime/src/iree/CMakeLists.txt
@@ -4,12 +4,4 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-add_subdirectory(base)
-add_subdirectory(builtins)
-add_subdirectory(hal)
-add_subdirectory(modules)
-add_subdirectory(runtime)
-add_subdirectory(schemas)
-add_subdirectory(task)
-add_subdirectory(testing)
-add_subdirectory(vm)
+iree_add_all_subdirs()
