+++ runtime/src/iree/testing/vulkan/iree-run-module-vulkan-gui-main.cc
@@ -16,7 +16,7 @@
 #include "iree/base/status_cc.h"
 #include "iree/hal/vulkan/registration/driver_module.h"
 #include "iree/modules/hal/module.h"
-#include "iree/tools/utils/vm_util.h"
+#include "iree/utils/vm_util.h"
 #include "iree/vm/api.h"
 #include "iree/vm/bytecode_module.h"
