+++ runtime/src/iree/utils/image_util.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/tools/utils/image_util.h"
+#include "iree/utils/image_util.h"
 
 #include <math.h>
