+++ runtime/src/iree/utils/trace_replay.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/tools/utils/trace_replay.h"
+#include "iree/utils/trace_replay.h"
 
 #include <ctype.h>
 #include <inttypes.h>
