+++ samples/iree_vision_inference/iree-run-mnist-module.c
@@ -11,7 +11,7 @@
 #include <float.h>
 
 #include "iree/runtime/api.h"
-#include "iree/tools/utils/image_util.h"
+#include "iree/utils/image_util.h"
 #include "iree_vision_inference/mnist_bytecode_module_c.h"
 
 iree_status_t Run(const iree_string_view_t image_path) {
